requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
google-generativeai==0.8.3
requests==2.32.3
